Allow codegen to be invoked as a library
